Move schema migrations into a submodule
