+++ dbt_platform_helper/providers/schema_migrator.py
@@ -1,0 +1,58 @@
+from collections import Counter
+from collections import OrderedDict
+from copy import deepcopy
+from typing import Protocol
+
+from dbt_platform_helper.platform_exception import PlatformException
+from dbt_platform_helper.providers.schema_migrations.schema_v0_to_v1_migration import (
+    SchemaV0ToV1Migration,
+)
+
+
+class InvalidMigrationConfigurationException(PlatformException):
+    pass
+
+
+class PlatformConfigSchemaMigrationException(PlatformException):
+    pass
+
+
+class SchemaMigrationProtocol(Protocol):
+    def from_version(self) -> int: ...
+
+    def migrate(self, platform_config: dict) -> dict: ...
+
+
+ALL_MIGRATIONS = [SchemaV0ToV1Migration()]
+
+
+class Migrator:
+    def __init__(self, *migrations: SchemaMigrationProtocol):
+        self.migrations = sorted(migrations, key=lambda m: m.from_version())
+        from_version_counts = Counter([migration.from_version() for migration in self.migrations])
+        duplicate_from_versions = [count for count in from_version_counts.values() if count > 1]
+
+        if duplicate_from_versions:
+            raise InvalidMigrationConfigurationException(
+                "`from_version` parameters must be unique amongst migrations"
+            )
+
+    def migrate(self, platform_config: dict) -> dict:
+        out = OrderedDict(deepcopy(platform_config))
+        if "schema_version" not in out:
+            out["schema_version"] = 0
+
+        if "default_versions" in out:
+            out.move_to_end("default_versions", last=False)
+        if "schema_version" in out:
+            out.move_to_end("schema_version", last=False)
+        if "application" in out:
+            out.move_to_end("application", last=False)
+
+        for migration in self.migrations:
+            if migration.from_version() != out["schema_version"]:
+                continue
+            out = migration.migrate(out)
+            out["schema_version"] += 1
+
+        return dict(out)
